Remove broken deployment cleanup step. (#45)
The cleanup step in the GitHub Action deployment should remove all
coa*.zip zipped executables from the GitHub release automatically so
they will only be available on Itch.io. The plugin is not working
because it says it can't find any assets to remove, so remove it for now
and as a workaround manually remove the assets from the GitHub release
using the web interface.